Feat/leaderboard (#141)
* add prompt syst

* add ruff info in contributong

* sort all def used in differents view in a utils file

* ruff

* change order column

* fix nb ligne of leaderboard

---------
