+++ eg1/api/crud.py
@@ -359,6 +359,7 @@ def get_leaderboard(
             models.Model.name.label("model_name"),
             models.Dataset.name.label("dataset_name"),
             main_metric_subquery.c.main_score.label("main_metric_score"),
+            models.Model.prompt_system,
             models.Model.sampling_params,
             models.Model.extra_params,
             models.Experiment.created_at.label("created_at"),
@@ -410,6 +411,7 @@ def get_leaderboard(
             dataset_name=result.dataset_name,
             main_metric_score=float(result.main_metric_score) if result.main_metric_score else None,
             other_metrics=other_metrics_dict,
+            prompt_system=result.prompt_system,
             sampling_param={k: str(v) for k, v in (result.sampling_params or {}).items()},
             extra_param={k: str(v) for k, v in (result.extra_params or {}).items()},
             created_at=result.created_at,
